# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TubePulse API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /keywords/search:
    get:
      operationId: searchKeywords
      tags: [keywords]
      summary: Search for keyword data
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
        - in: query
          name: channelId
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Keyword search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeywordSearchResult"
  /keywords/suggestions:
    get:
      operationId: getKeywordSuggestions
      tags: [keywords]
      summary: Get tag suggestions for a video
      parameters:
        - in: query
          name: videoId
          required: true
          schema:
            type: string
        - in: query
          name: channelId
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Tag suggestions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KeywordSuggestion"
  /keywords/rankings:
    get:
      operationId: getKeywordRankings
      tags: [keywords]
      summary: Get tag ranking history for a video
      parameters:
        - in: query
          name: channelId
          required: true
          schema:
            type: integer
        - in: query
          name: videoId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ranking history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KeywordRanking"
...
components:
  schemas:
...
    KeywordSearchResult:
      type: object
      required: [query, keywords]
      properties:
        query:
          type: string
        keywords:
          type: array
          items:
            $ref: "#/components/schemas/KeywordData"
    KeywordData:
      type: object
      required: [query, searchVolume, competitionScore, overallScore, relatedKeywords, trendData]
      properties:
        query:
          type: string
        searchVolume:
          type: string
          enum: [very_low, low, medium, high, very_high]
        competitionScore:
          type: integer
        overallScore:
          type: integer
        relatedKeywords:
          type: array
          items:
            type: string
        trendData:
          type: array
          items:
            type: number
    KeywordSuggestion:
      type: object
      required: [tag, score, searchVolume]
      properties:
        tag:
          type: string
        score:
          type: integer
        searchVolume:
          type: string
          enum: [very_low, low, medium, high, very_high]
    KeywordRanking:
      type: object
      required: [tag, rank, date]
      properties:
        tag:
          type: string
        rank:
          type: ["integer", "null"]
        date:
          type: string
          format: date
...